refactor: Async ThreadPool설정 클래스 위치 이동
